use custom format when relaying from syslog_aggregator
The rsyslogd default format resulted in "localhost" as the hostname.
Ops was using the hostname to determine where to store the log data,
relying on being able to resolve a fqdn to determine the network
(i.e. dev vs. staging).  We settled on using "cf-<IP address>"
as the hostname, so they could use regex rules to match and store
appropriately.

Additionally, I extended the format to include the IP address that
the logging data originated from, so that logs from different vms
running the same job (e.g. deas) could be easily distinguished
from each other.

Change-Id: I45017b16064aec8a364eb0d6a712a4c3a12130df