Add oauth2 service as a new job
This service binds OAuth2 client credentials to a user
app in VCAP_SERVICES. They can then be used by the app
to identify and authenticate users and operate the
cloud_controller on their behalf (i.e. no more apps
collecting user credentials).

There is a test case already in yeti.

[Fixes #39342839] [cfid-307] Create job in release repo
for UAA as a service

Change-Id: Ied944b9b0cb194bee32d76acd3c62d0cf170ebe1