Allow login url to be specified in cc
Explanation: the UAA used to do everything, then we broke
out authentication into the Login Server quite some time ago.
For backwards compatibility we kept it in the UAA, but we
need to retire that feature.  And the cloud controller needs
to point at the Login Server (if there is one) for anything
that needs user authentication (including vmc).