Document login.links.xyz properties and defaults
Nowhere in cf-release is it mentioned that the 3 available login.links.xyz
properties are and what their default values are. This commit brings across
the current default values that login-server itself provides if
login.links.home or login.links.passwd or login.links.signup are not provided.

They are commented out so as to not force properties.login to be set
if the deployment file doesn't want it.