don't write the CC's pidfile in the monit ctl
This led to the pid being written before migrations could even run.
Which is bad. Because it will look like it's up as the migrations are
running, and a deployment will continue to the rest of the CCs, who
expect the migrations to have completed.

The CC already writes the same pidfile on startup, as it's in its config
file. This was probably just mistakenly copy-pasted before.

[finishes #57704354]