Fix NFS stale file handler
If Cloud Controller is using debian nfs server as local droplets storage
and the nfs server disk is resized the file handle on the client is
stale until the client is restarted. HYPOTHESIS: because the disk
re-size involves copying the NFS served files to a new partition which
is then remounted and re-served from the same NFS mount, the client
retains (inode?) a reference to the original file but is unable to
reconnect to the copied version on the new partition (at the same path
location).

Signed-off-by: Aram Price <aprice@pivotallabs.com>