Explicit release name for each job
This means that when adding an additional release to a resulting
manifest, the person does not need to manually add "release: cf"
to each and every job.

There is no affect to people who do not add additional releases
to their CF deployment file.

Allow job release name to be specified by meta.release.name