Java Buildpack 2.1
Starting with the 2.1 release of the Java Buildpack, packaged versions
of the buildpack should not be linked via submodule.  Instead, a
packaged version of the buildpack, (complete with version baked into
the code) should be used.

To facilitate both Cloud Foundry open source and PivotalCF, there are
two different versions of the buildpack.  The first, online buildpack,
should be used in environments that have Internet access (e.g.
https://run.pivotal.io).  The second, offline buildpack, should be
used in environments that do not have Internet access (e.g.
PivotalCF).

This change removes the existing Java Buildpack submodule and replaces
it with a reference to the packaged buildpack in the blobstore.  This
packaging style was based on the gloang packaging that already
existed.