Switch login job to look at uaa.login.client_secret only
This allows removal of the uaa.clients.login.secret from the login job spec,
thereby allowing the default UAA login client definition [1] to take place, if
the login client isn't present in the manifest, making login similar to other
default uaa clients.

When uaa.clients.login.secret is present in the login job spec, colocating UAA
and login on the same machine, and not specifying uaa.clients.login results in
an incorrect configuration for the login client in the uaa configuration.

This also allows for removal of clients that are specific to Pivotal's deploy
from deployment manifests without breaking default clients.

[1] https://github.com/cloudfoundry/cf-release/blob/master/jobs/uaa/templates/uaa.yml.erb#L80-L89