Enable logging and fix syntax errors
Currently, this script errors out with a syntax error since "mount_nfs()" is syntax for creating a function, not invoking one. The same problem exists with "unmount_nfs()". Also, the "start)" case wasn't terminated.

Additionally, this job should log its output like the other jobs in CF so I added the inclusion of comm/utils.sh which enables this.