DNS host names cannot contain an underscore
The current consul templates assume that the service name matches the
job name. This causes problems for many bosh jobs that use an underscore
in the name since some ruby clients that do URI validation.  Since DNS
does not support host names that contain underscores, ruby gets upset.

This is a tactical change to convert underscores to hyphens for the time
being.

[#86242828]