Enable overwriting of property uaa.require_https
The uaa job defines a property require_https, which could be
used to configure the uaa to accept only https requests (plaing
http requests are then redirected to https).

This patch adds the require_https property to cf-properties.yml,
so that it could be overwritten in a configuration stub file.
The default is set to 'false' to keep the current default
behaviour of the uaa job if the property is missing.