Preserve rootfs permissions
Move the extraction of the cflinuxfs2 rootfs into the dea_next job vm. This
will ensure that the ownership of /home/vcap dir in the rootfs is vcap:vcap.
The compilation & distribution steps do not preserve the ownership and that's
why we are moving the extraction of the rootfs to the warden_ctl script.

Signed-off-by: Svett Ralchev <sralchev@pivotal.io>