Allow override NFS configuration in stub
The default NFS server address is defined by the static IP of the `nfs_z1` job,
which forces define static IPs. Additionally, the default allowed NFS clients are 
defined by the network subnets ranges, which forces having to such ranges. 

This is problematic with some CPIs which do not support static IPs or subnets.

This commit adds a `merge` for the NFS server address and `allow_from_entries`

A new value `meta.nfs_client_ranges` is added to allow the stub to define 
different number of child entries under `allow_client_ranges`. It is define
outside of `meta.nfs_server` to avoid pollute the generated manifest.