Allow override default network for router jobs
Allow override the network default name and configuration (`cf1` or
`cf2`) for the `router_z1` and `router_z2` jobs and their corresponding
resource pools.

This feature is required in the case that a different
network is needed to configure the routers' loadbalancer.
For instance,  in the case of the [BOSH Google
CPI](https://github.com/frodenas/bosh-google-cpi), the automatic
loadbalancer registering requires the definition of alternate networks
with a custom `cloud_properties.target_pool` entry.