Replace mysql client with libmariadb client
We had to modify the vanilla mysql_config for
mariadb-5.5/libmariadbclient-dev_5.5.36-1_amd64.deb to build the mysql2
gem. We needed to replace all lib/* paths with lib/x86_64-linux-gnu or
else the package would not compile. We stole this change from the 5.5.46
ubuntu version. This change is in the uploaded blob.

[#110671586]

Signed-off-by: Simon Leung <leungs@us.ibm.com>