Add ruby-2.3 package (using Ruby 2.3.1)
This package is specifically _not_ name ruby-2.3.1 to make it easier to
update 2.3.x as needed. There may need to be some coordination between
teams that are using 2.3.x if there are failures with the new release.
Right now, only CAPI is using 2.3, and only runtime-og is using another
Ruby, so the chances of this actually being a problem are small.

[#110855966]

Signed-off-by: Dan Wendorf <dwendorf@pivotal.io>