Allow configuration of artifacts directory
Add some logic to check if the `smoke_tests.artifacts_directory` property
is set and add it to the smoke test configuration. This makes it possible
to turn on logging of CF_TRACE output from the manifest (with the intent
of shipping the logs to another location, which will be useful for debugging
failed tests).

Once the template is rendered you will be left with some fairly hideous
whitespace (a linefeed before and after the artifacts directory key). I
did not implement trimming of this because the Bosh repository shows
inconsistent support for the trim_mode parameters[1] in ERB.

The artifacts directory is hard coded for the acceptance tests job[2]. Other
than a mild inconsistency there is no reason to change this behaviour.

[1] http://ruby-doc.org/stdlib-2.3.1/libdoc/erb/rdoc/ERB.html#method-c-new
[2] https://github.com/cloudfoundry/cf-release/blob/v245/jobs/acceptance-tests/templates/config.json.erb#L24