Move clock_global into clock_z{1,2} to enable HA mode
This requires https://www.pivotaltracker.com/story/show/136621261 to be
released and merged into cf-release before it can be merged.

The clock job will hold onto a lock in the CCDB to ensure only one clock
is scheduling jobs at a time, which enables deploying more than one
instance of it.

Signed-off-by: Luan Santos <lsantos@pivotal.io>