Revert "Bump consul-release to v162"
This bump is breaking both routing deployments and cf-mysql deployments
becasue those releases attempt to consume the consul link. We're getting
the standard multi-AZ-jobs-link problem.

We need to first get a version of cf-mysql-release and routing-release
that nil our the consul link in their manifest generation in order to
bump to consul.

This reverts commit eb366c455145da725e7ce2531ad87137d4dec357.