Remove minimal AWS
This manifest used to provide two useful things:
1. A basic deployment with hard-coded values that made it easy for
first-time deployers to try out CF. We now have cf-deployment
(https://github.com/cloudfoundry/cf-deployment) which is better at this.
2. Alerted the RelInt PM to backwards-incompatible changes in
cf-release, so that release notes could be explain the update. It's been
many months since that kind of value, and we're moving to deprecate
cf-release as soon as possible, so we're not getting this value any
longer either.